+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Made the SC/h line on hud hidden when its rate is zero
 
 ### Fixes
+- Disabled Party Finder hover effects while using the peek keybind
 - Fixed sea creature glow only displaying one entity and invisible entities sometimes
 - Fixed bossbar and merge health settings in /rfuscedit not properly applying to sea creatures
 - Peek party finder keybind now closes the creation window / presets